Get BlockTag tooltips working again.
